==> Code/classification/04-domains.ipynb <==
# Domains

import pandas as pd
import numpy as np

tweets = pd.read_csv('data/porn/tweets.csv', encoding='utf-8-sig', sep='\t')

tweets['url'].isnull().value_counts()

users.columns

tweets.columns

tweets['url']

from urllib.parse import urlparse

def extract_domain(domain):
    data = urlparse(string)
    return data.netloc

string

extract_domain(string)
