==> Code/classification/05-test_random_user.ipynb <==
import pandas as pd
import numpy as np
import tweepy
from sklearn.ensemble import RandomForestClassifier

### targets:
####    0: porn
####    1: propaganda
####    2: spam
####    3: fake followers
####    4: genuine accounts

# Test users
### 3201005265 - porn - 0
### 3593771535 - porn - 0
### 126523026 - spam - 2
### 3318703632 - propaganda - 1
### 875124919 - propaganda - 1

# create connection with Twitter API

CONSUMER_KEY = 'J3uhwWHdmSt3uD69ry8r2kc3B'
CONSUMER_SECRET = 'LeuM7dVfFbXH0bc5fFkZIxlNiUXyjIx4Kcjh58HeWs2TrLijDx'
ACCESS_TOKEN = '327497511-5dNjiGaTQHASljoxNr1qlMOlHrrSB21HBFBDjx2E'
ACCESS_TOKEN_SECRET = 'tHLqwlhHYJWptW3femyq0rbMW6ZItu5yLQ3DFGJqJ8Xlg'

auth = tweepy.OAuthHandler(CONSUMER_KEY, CONSUMER_SECRET)
auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)
api = tweepy.API(auth)

ids = [3201005265, 3593771535, 126523026, 3318703632, 875124919]

test_set = []
for elem in ids:
    test_set.append(api.get_user(elem))

df = pd.DataFrame(columns=["statuses_count","followers_count","friends_count","favourites_count","listed_count","url","default_profile_image","profile_use_background_image","profile_background_image_url_https","profile_background_tile","profile_background_image_url","description"])
for user in test_set:
    df2 = pd.DataFrame([[user.statuses_count,
                            user.followers_count,
                            user.friends_count,
                            user.favourites_count,
                            user.listed_count,
                            user.url,
                            user.default_profile_image,
                            user.profile_use_background_image,
                            user.profile_background_image_url_https,
                            user.profile_background_tile,
                            user.profile_background_image_url,
                            user.description]],
                            columns=["statuses_count","followers_count","friends_count","favourites_count","listed_count","url","default_profile_image","profile_use_background_image","profile_background_image_url_https","profile_background_tile","profile_background_image_url","description"]
                          )
    df = df.append(df2, ignore_index=True)

df['description'] = df['description'].fillna('')
df['description'] = df['description'].apply(lambda x: len(x))

df['url']

df['default_profile_image'] = df['default_profile_image'].astype(int)

df['profile_background_tile'] = df['profile_background_tile'].astype(int)

df['profile_use_background_image'] = df['profile_use_background_image'].astype(int)

df['profile_background_image_url'] = df['profile_background_image_url'].apply(lambda x: int(not(int(pd.isnull(x)))))

df['profile_background_image_url_https'] = df['profile_background_image_url_https'].apply(lambda x: int(not(int(pd.isnull(x)))))

df['url'] = df['url'].apply(lambda x: int(not(int(pd.isnull(x)))))

df

train_set = pd.read_csv('data/full/baseline_dataset.csv', encoding='utf-8-sig')

train_set = train_set.drop('Unnamed: 0', axis=1)

df = df.reindex_axis(sorted(df.columns), axis=1)

df.to_csv('data/full/baseline_testset.csv')

target = 'target'
features = train_set.columns[train_set.columns!=target]

X = train_set[features]
y = train_set[target]

# Fitting Random Forest Classification to the Training set
forest = RandomForestClassifier(n_estimators = 10, criterion = 'entropy', random_state = 42)
forest.fit(X, y)

df

train_set
