==> Code/classification/06-uniform_data.ipynb <==
import pandas as pd
import numpy as np
from IPython.display import clear_output

## Users

# users
porn = pd.read_csv('data/porn/users.csv')
spam = pd.read_csv('data/spam/users.csv')
ff = pd.read_csv('data/fake_followers/users.csv')
news = pd.read_csv('data/propaganda/users.csv')
genuine = pd.read_csv('data/genuine/users.csv')

porn.columns

ff.columns

ff.columns[np.in1d(ff.columns, porn.columns, invert=True)]

news.columns

genuine.columns

porn.drop(columns=['Unnamed: 0'], inplace=True)

len(porn.columns)

spam.drop(columns=['Unnamed: 0'], inplace=True)

len(spam.columns)

ff.drop(columns=['Unnamed: 0', 'Unnamed: 0.1', 'profile_banner_url', 'updated'], inplace=True)

len(ff.columns)

len(news.columns)

genuine.columns[np.in1d(genuine.columns, porn.columns, invert=True)]

genuine.drop(columns=['profile_banner_url', 'timestamp', 'crawled_at', 'updated', 'test_set_1' ,'test_set_2'], inplace=True)

len(genuine.columns)

## integro con varol

varol = pd.read_csv('data/varol/users.csv')

varol['target'].value_counts()

porn_v = varol[varol.target == 0]
porn_v.drop(columns=['target'], inplace=True)
porn = pd.concat([porn, porn_v])

news_v = varol[varol.target == 1]
news_v.drop(columns=['target'], inplace=True)
news = pd.concat([news, news_v])

spam_v = varol[varol.target == 2]
spam_v.drop(columns=['target'], inplace=True)
spam = pd.concat([spam, spam_v])

ff_v = varol[varol.target == 3]
ff_v.drop(columns=['target'], inplace=True)
ff = pd.concat([ff, ff_v])

genuine_v = varol[varol.target == 4]
genuine_v2 = varol[varol.target == 5]
genuine_v.drop(columns=['target'], inplace=True)
genuine_v2.drop(columns=['target'], inplace=True)
genuine = pd.concat([genuine, genuine_v, genuine_v2])

### save

porn.to_csv('data/porn/users.csv', index=False)
news.to_csv('data/propaganda/users.csv', index=False)
ff.to_csv('data/fake_followers/users.csv', index=False)
spam.to_csv('data/spam/users.csv', index=False)
genuine.to_csv('data/genuine/users.csv', index=False)

ff2 = pd.read_csv('data/fake_followers/dibe_users.csv')

ff.columns

ff2.columns

np.in1d(ff.columns, ff2.columns)

np.in1d(ff2.columns, ff.columns)

ff.shape

ff2.shape

ff = pd.concat([ff2, ff])

ff.to_csv('data/fake_followers/users.csv', index=False)

## Tweets

varol = pd.read_csv('data/varol/tweets.csv', sep='\t')

users = pd.read_csv('data/varol/users.csv')

cont = 0
targets = []
for tweet in varol.itertuples():
    cont += 1
    clear_output()
    print(str(cont) + ' / 65873')
    user = int(tweet.user_id)
    print(user)
    target = int(users[users.id == user].target.item())
    targets.append(target)

varol['target'] = targets

varol.target.value_counts()

varol.drop(columns=['withheld_copyright','withheld_in_countries', 'withheld_scope'], inplace=True)

### Porn

porn = pd.read_csv('data/porn/tweets.csv', sep = '\t')

porn.columns

porn.drop(columns=['Unnamed: 0'], inplace=True)

varol.columns

np.in1d(varol.columns, porn.columns)

np.in1d(varol.columns, porn.columns)

porn2 = varol[varol.target == 0]

porn2.drop(columns=['target'], inplace=True)

porn = pd.concat([porn, porn2])

porn.to_csv('data/porn/tweets.csv', index=False, sep='\t', encoding='utf-8-sig')

del porn, porn2

### Spam

spam = pd.read_csv('data/spam/tweets.csv', sep = '\t')

spam.columns

spam.drop(columns=['Unnamed: 0'], inplace=True)

np.in1d(varol.columns, spam.columns)

spam2 = varol[varol.target == 2]
spam2.drop(columns=['target'], inplace=True)

spam = pd.concat([spam, spam2])

spam.to_csv('data/spam/tweets.csv', index=False, sep='\t', encoding='utf-8-sig')

del spam, spam2

### ff

ff = pd.read_csv('data/fake_followers/tweets.csv', sep = '\t')

ff.columns

np.in1d(ff.columns, varol.columns)

ff2 = varol[varol.target == 3]
ff2.drop(columns=['target'], inplace=True)

ff = pd.concat([ff, ff2])

ff.to_csv('data/fake_followers/tweets.csv', index=False, sep='\t', encoding='utf-8-sig')

del ff, ff2

ff2 = pd.read_csv('data/fake_followers/dibe_tweets.csv', sep = '\t')

np.in1d(ff2.columns, ff.columns)

ff2.columns

ff2.drop(columns=['withheld_copyright','withheld_in_countries', 'withheld_scope'], inplace=True)

ff = pd.concat([ff2, ff])

ff.to_csv('data/fake_followers/tweets.csv', index=False, sep='\t', encoding='utf-8-sig')

### News

news = pd.read_csv('data/propaganda/tweets.csv', sep = '\t')

len(news)

news.columns

np.in1d(news.columns, varol.columns)

np.in1d(varol.columns, news.columns)

news2 = varol[varol.target == 1]
news2.drop(columns=['target'], inplace=True)

news = pd.concat([news, news2])

len(news)

news.drop(columns=['withheld_in_countries'], inplace=True)

news.to_csv('data/propaganda/tweets.csv', index=False, sep='\t', encoding='utf-8-sig')

del news, news2

### Genuine

gen = pd.read_csv('data/genuine/tweets.csv')

gen.columns

del gen
