==> Code/classification/07-features_vector-Copy1.ipynb <==
# Features Vector

import pandas as pd
import numpy as np
from IPython.display import clear_output

### Load users features

porn_users = pd.read_csv('data/porn/users.csv', encoding='utf-8-sig')
prop_users = pd.read_csv('data/propaganda/users.csv', encoding='utf-8-sig')
spam_users = pd.read_csv('data/spam/users.csv', encoding='utf-8-sig')
fake_users = pd.read_csv('data/fake_followers/users.csv', encoding='utf-8-sig')
genuine_users = pd.read_csv('data/genuine/users.csv', encoding='utf-8-sig')

### Load tweets features

#Porn
porn_tweet_intrad = pd.read_csv('data/porn/tweet_intradistance.csv', encoding='utf-8-sig')
porn_url_intrad = pd.read_csv('data/porn/url_intradistance.csv', encoding='utf-8-sig')
porn_context_score = pd.read_csv('data/porn/context_score.csv', encoding='utf-8-sig')
porn_tweet_descr = pd.read_csv('data/porn/descriptive_features.csv', encoding='utf-8-sig')

#Propaganda
prop_tweet_intrad = pd.read_csv('data/propaganda/tweet_intradistance.csv', encoding='utf-8-sig')
prop_url_intrad = pd.read_csv('data/propaganda/url_intradistance.csv', encoding='utf-8-sig')
prop_context_score = pd.read_csv('data/propaganda/context_score.csv', encoding='utf-8-sig')
prop_tweet_descr = pd.read_csv('data/propaganda/descriptive_features.csv', encoding='utf-8-sig')

#Spam
spam_tweet_intrad = pd.read_csv('data/spam/tweet_intradistance.csv', encoding='utf-8-sig')
spam_url_intrad = pd.read_csv('data/spam/url_intradistance.csv', encoding='utf-8-sig')
spam_context_score = pd.read_csv('data/spam/context_score.csv', encoding='utf-8-sig')
spam_tweet_descr = pd.read_csv('data/spam/descriptive_features.csv', encoding='utf-8-sig')

#Fake followers
fake_tweet_intrad = pd.read_csv('data/fake_followers/tweet_intradistance.csv', encoding='utf-8-sig')
fake_url_intrad = pd.read_csv('data/fake_followers/url_intradistance.csv', encoding='utf-8-sig')
fake_context_score = pd.read_csv('data/fake_followers/context_score.csv', encoding='utf-8-sig')
fake_tweet_descr = pd.read_csv('data/fake_followers/descriptive_features.csv', encoding='utf-8-sig')

#Genuine
gen_tweet_intrad = pd.read_csv('data/genuine/tweet_intradistance.csv', encoding='utf-8-sig')
gen_url_intrad = pd.read_csv('data/genuine/url_intradistance.csv', encoding='utf-8-sig')
gen_context_score = pd.read_csv('data/genuine/context_score.csv', encoding='utf-8-sig')
gen_tweet_descr = pd.read_csv('data/genuine/descriptive_features.csv', encoding='utf-8-sig')

## Drop Duplicates

#Porn
porn_tweet_intrad = porn_tweet_intrad.drop_duplicates()
print(len(porn_tweet_intrad))
porn_url_intrad = porn_url_intrad.drop_duplicates()
print(len(porn_url_intrad))
porn_context_score = porn_context_score.drop_duplicates()
print(len(porn_context_score))
porn_tweet_descr = porn_tweet_descr.drop_duplicates()
print(len(porn_tweet_descr))
porn_users = porn_users.drop_duplicates()
print(len(porn_users))

#Cut useless users
ids = porn_users.id.drop_duplicates()
porn_users = porn_users.iloc[ids.index]
print(len(porn_users))

#Propaganda
prop_tweet_intrad = prop_tweet_intrad.drop_duplicates()
print(len(prop_tweet_intrad))
prop_url_intrad = prop_url_intrad.drop_duplicates()
print(len(prop_url_intrad))
prop_context_score = prop_context_score.drop_duplicates()
print(len(prop_context_score))
prop_tweet_descr = prop_tweet_descr.drop_duplicates()
print(len(prop_tweet_descr))
prop_users = prop_users.drop_duplicates()
print(len(prop_users))

#Spam
spam_tweet_intrad = spam_tweet_intrad.drop_duplicates()
print(len(spam_tweet_intrad))
spam_url_intrad = spam_url_intrad.drop_duplicates()
print(len(spam_url_intrad))
spam_context_score = spam_context_score.drop_duplicates()
print(len(spam_context_score))
spam_tweet_descr = spam_tweet_descr.drop_duplicates()
print(len(spam_tweet_descr))
spam_users = spam_users.drop_duplicates()
print(len(spam_users))

#Fake followers
fake_tweet_intrad = fake_tweet_intrad.drop_duplicates()
print(len(fake_tweet_intrad))
fake_url_intrad = fake_url_intrad.drop_duplicates()
print(len(fake_url_intrad))
fake_context_score = fake_context_score.drop_duplicates()
print(len(fake_context_score))
fake_tweet_descr = fake_tweet_descr.drop_duplicates()
print(len(fake_tweet_descr))
fake_users = fake_users.drop_duplicates()
print(len(fake_users))

#Cut useless users
ids = fake_users.id.drop_duplicates()
fake_users = fake_users.loc[ids.index]
print(len(fake_users))

#Genuine
gen_tweet_intrad = gen_tweet_intrad.drop_duplicates()
print(len(gen_tweet_intrad))
gen_url_intrad = gen_url_intrad.drop_duplicates()
print(len(gen_url_intrad))
gen_context_score = gen_context_score.drop_duplicates()
print(len(gen_context_score))
gen_tweet_descr = gen_tweet_descr.drop_duplicates()
print(len(gen_tweet_descr))
gen_users = genuine_users.drop_duplicates()
print(len(gen_users))

## Merge dataframes

#Porn
porn_1 = pd.merge(porn_tweet_intrad, porn_url_intrad, on='user_id')
porn_2 = pd.merge(porn_context_score, porn_tweet_descr, on='user_id')
porn_df = pd.merge(porn_1, porn_2, on='user_id')

#porn_users.rename(columns={'id': 'user_id'}, inplace=True)
#porn_df = pd.merge(porn_users, porn_df, on='user_id')

porn_df.columns

#Propaganda
prop_1 = pd.merge(prop_tweet_intrad, prop_url_intrad, on='user_id')
prop_2 = pd.merge(prop_context_score, prop_tweet_descr, on='user_id')
prop_df = pd.merge(prop_1, prop_2, on='user_id')

prop_users.rename(columns={'id': 'user_id'}, inplace=True)
prop_df = pd.merge(prop_users, prop_df, on='user_id')

#Spam
spam_1 = pd.merge(spam_tweet_intrad, spam_url_intrad, on='user_id')
spam_2 = pd.merge(spam_context_score, spam_tweet_descr, on='user_id')
spam_df = pd.merge(spam_1, spam_2, on='user_id')

spam_users.rename(columns={'id': 'user_id'}, inplace=True)
spam_df = pd.merge(spam_users, spam_df, on='user_id')

#Fake followers
fake_1 = pd.merge(fake_tweet_intrad, fake_url_intrad, on='user_id')
print(len(fake_1))
fake_2 = pd.merge(fake_context_score, fake_tweet_descr, on='user_id')
print(len(fake_1))
fake_df = pd.merge(fake_1, fake_2, on='user_id')
print(len(fake_df))

fake_users.rename(columns={'id': 'user_id'}, inplace=True)
fake_df = pd.merge(fake_users, fake_df, on='user_id')

print(len(porn_df))

np.in1d(fake_tweet_intrad.user_id, fake_url_intrad.user_id, invert=True).sum()

#Genuine
gen_1 = pd.merge(gen_tweet_intrad, gen_url_intrad, on='user_id')
gen_2 = pd.merge(gen_context_score, gen_tweet_descr, on='user_id')
gen_df = pd.merge(gen_1, gen_2, on='user_id')

genuine_users.rename(columns={'id': 'user_id'}, inplace=True)
gen_df = pd.merge(genuine_users, gen_df, on='user_id')

porn_df.to_csv('data/porn/train.csv', index=False)
prop_df.to_csv('data/propaganda/train.csv', index=False)
spam_df.to_csv('data/spam/train.csv', index=False)
fake_df.to_csv('data/fake_followers/train.csv', index=False)
gen_df.to_csv('data/genuine/train.csv', index=False)

## Controllo integrità

porn_df = pd.read_csv('data/porn/train.csv')
news_df = pd.read_csv('data/propaganda/train.csv')
spam_df = pd.read_csv('data/spam/train.csv')
fake_df = pd.read_csv('data/fake_followers/train.csv')
gen_df = pd.read_csv('data/genuine/train.csv')

len(porn_df)

len(set(porn_df.user_id))

len(fake_df)

len(set(fake_df.user_id))

len(news_df)

len(set(news_df.user_id))

len(spam_df)

len(set(spam_df.user_id))

len(gen_df)

len(set(gen_df.user_id))
