==> Code/classification/07-tweets_categories.ipynb <==
import pandas as pd
from googletrans import Translator
from uclassify import uclassify
import json
import urllib3

df = pd.read_csv('data/porn/tweets.csv', sep = '\t')

df = df[['id','full_text']]

len(df)

df2 = pd.read_csv('data/genuine/tweets.csv', sep='\t')
df2 = df2[['id','full_text']]

df = pd.concat([df,df2],ignore_index=True)

del df2

len(df)

df2 = pd.read_csv('data/propaganda/tweets.csv', sep='\t')
df2 = df2[['id','full_text']]
df = pd.concat([df,df2],ignore_index=True)
del df2

len(df)

df2 = pd.read_csv('data/spam/tweets.csv', sep='\t')
df2 = df2[['id','full_text']]
df = pd.concat([df,df2],ignore_index=True)
del df2

len(df)

df2 = pd.read_csv('data/fake_followers/tweets.csv', sep='\t')
df2 = df2[['id','full_text']]
df = pd.concat([df,df2],ignore_index=True)
del df2

len(df)

df.to_csv('data/tweets_cat/id_full_text.csv', sep = '\t', index=False)

# Ricomincia da Qui

import pandas as pd
from googletrans import Translator
import requests
import re

df = pd.read_csv('data/tweets_cat/id_full_text.csv', sep = '\t')

df2 = df.head(1)

categories = []

for x in df2.itertuples():
    txt = x.full_text
    txt = re.sub(r'[^\w\d\s\,\.\:\@\#\/]+', '', txt)
    #txt = Translator.translate(text=txt).text
    response = requests.post('https://api.uclassify.com/v1/uclassify/iab-taxonomy-v2/classify', \
                             data = "{\"texts\": "+ txt) + "}", \
                            headers = "{'Authorization': 'Token ' + 'hL46Q29XYTSf'}")
    categories.append(response.content)

categories
