==> Code/classification/08-full_ds_creator.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix
from sklearn.externals import joblib
from sklearn import model_selection, linear_model
from sklearn.model_selection import StratifiedKFold
from sklearn.neighbors import KNeighborsClassifier
from sklearn import svm

porn = pd.read_csv('data/porn/train.csv') #0
news = pd.read_csv('data/propaganda/train.csv') #1
spam = pd.read_csv('data/spam/train.csv') #2
ff = pd.read_csv('data/fake_followers/train.csv') #3
genuine = pd.read_csv('data/genuine/train.csv') #4

porn['target'] = 0
news['target'] = 1
spam['target'] = 2
ff['target'] = 3
genuine['target'] = 4

full = pd.concat([porn,news,spam,ff,genuine], sort=True)

full.to_csv('data/full/train.csv', index=False)
