==> Code/classification/10-NB_train.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
import re
from sklearn.pipeline import Pipeline
from sklearn.model_selection import cross_val_score
import nltk
from nltk.stem.snowball import SnowballStemmer
from sklearn.metrics import f1_score
from nltk.corpus import stopwords
import pickle

users = pd.read_csv('data/full/train.csv')

users = users[['user_id','target']]

tweets = pd.concat([pd.read_csv('data/porn/tweets.csv', sep='\t')[['user_id','full_text']], \
          pd.read_csv('data/propaganda/tweets.csv', sep='\t')[['user_id','full_text']], \
          pd.read_csv('data/spam/tweets.csv', sep='\t')[['user_id','full_text']], \
          pd.read_csv('data/fake_followers/tweets.csv', sep='\t')[['user_id','full_text']]])

tweets = tweets.drop_duplicates()

tweets =  tweets.join(users.set_index('user_id'), on='user_id')

tweets = tweets[tweets.target!=4]

tweets.target.value_counts()

tweets.drop(['user_id'], axis=1, inplace=True)

tweets = tweets[~np.isnan(tweets.target)]

tweets.target = tweets.target.astype(int)

tweets

def remove_rt(x):
    if 'RT @' in x:
        try:
            return x[x.find(':')+2:]
        except:
            return x
    else:
        return x

#def translate(x):
#    try:
#        return translator.translate(x).text
#    except:
#        return x

stop_words = stopwords.words('english')

def remove_stop(x):
    return [word for word in x.split() if word not in stop_words]

tweets['full_text'] = tweets['full_text'].apply(lambda x: remove_rt(x))

tweets['full_text'] = tweets['full_text'].apply(lambda x: re.sub(r'http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+', '', x))

tweets['full_text'] = tweets['full_text'].apply(lambda x: re.sub(r'[^\w\s]','',x))

tweets['full_text'] = tweets['full_text'].apply(lambda x: x.lower())

tweets['full_text'] = tweets['full_text'].apply(lambda x: remove_stop(x))

tweets['full_text'] = tweets['full_text'].astype(str)

tweets = tweets[tweets.full_text!='[]']

tweets.target.value_counts()

stemmer = SnowballStemmer("english", ignore_stopwords=True)

class StemmedCountVectorizer(CountVectorizer):
    def build_analyzer(self):
        analyzer = super(CountVectorizer, self).build_analyzer()
        return lambda doc:(stemmer.stem(w) for w in analyzer(doc))


stem_vectorizer = StemmedCountVectorizer(stemmer)

pipeline = Pipeline([('vect', stem_vectorizer), 
                     ('tfidf', TfidfTransformer()), 
                     ('mnb', MultinomialNB(fit_prior=False))])
X_train, X_test, y_train, y_test = train_test_split(tweets.full_text, tweets.target, test_size=0.33, random_state=42)

clf = pipeline.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print('accuracy: ' + str(np.mean(y_pred == y_test)))
print('partial f1: ' + str(f1_score(y_test, y_pred, average=None)))
print('total f1: ' + str(f1_score(y_test, y_pred, average='macro')))

clf.predict_proba(pd.Series("try this awesome app"))

clf = pipeline.fit(tweets.full_text, tweets.target)

y_pred = clf.predict(tweets.full_text)

print('accuracy: ' + str(np.mean(y_pred == tweets.target)))
print('partial f1: ' + str(f1_score(tweets.target, y_pred, average=None)))
print('total f1: ' + str(f1_score(tweets.target, y_pred, average='macro')))

pickle.dump( clf, open( "../script2/models/nb.model", "wb" ) )
