==> Code/classification/16-bot_not_features_vector.ipynb <==
# Bot or Not Features Vector

import pandas as pd
import numpy as np
from IPython.display import clear_output

### Load users features

bot_users = pd.read_csv('data/bot_or_not/bot/users.csv', encoding='utf-8-sig')
gen_users = pd.read_csv('data/bot_or_not/gen/users.csv', encoding='utf-8-sig')

### Load tweets features

#Bot
porn_tweet_intrad = pd.read_csv('data/bot_or_not/bot/tweet_intradistance.csv', encoding='utf-8-sig')
porn_url_intrad = pd.read_csv('data/bot_or_not/bot/url_intradistance.csv', encoding='utf-8-sig')
porn_tweet_descr = pd.read_csv('data/bot_or_not/bot/descriptive_features.csv', encoding='utf-8-sig')

#Gen
prop_tweet_intrad = pd.read_csv('data/bot_or_not/gen/tweet_intradistance.csv', encoding='utf-8-sig')
prop_url_intrad = pd.read_csv('data/bot_or_not/gen/url_intradistance.csv', encoding='utf-8-sig')
prop_tweet_descr = pd.read_csv('data/bot_or_not/gen/descriptive_features.csv', encoding='utf-8-sig')

## Drop Duplicates

#Porn
porn_tweet_intrad = porn_tweet_intrad.drop_duplicates()
print(len(porn_tweet_intrad))
porn_url_intrad = porn_url_intrad.drop_duplicates()
print(len(porn_url_intrad))
porn_tweet_descr = porn_tweet_descr.drop_duplicates()
print(len(porn_tweet_descr))
bot_users = bot_users.drop_duplicates()
print(len(bot_users))

#Propaganda
prop_tweet_intrad = prop_tweet_intrad.drop_duplicates()
print(len(prop_tweet_intrad))
prop_url_intrad = prop_url_intrad.drop_duplicates()
print(len(prop_url_intrad))
prop_tweet_descr = prop_tweet_descr.drop_duplicates()
print(len(prop_tweet_descr))
gen_users = gen_users.drop_duplicates()
print(len(gen_users))

## Merge dataframes

#Porn
porn_1 = pd.merge(porn_tweet_intrad, porn_url_intrad, on='user_id')
porn_df = pd.merge(porn_1, porn_tweet_descr, on='user_id')

#porn_users.rename(columns={'id': 'user_id'}, inplace=True)
#porn_df = pd.merge(porn_users, porn_df, on='user_id')

bot_users.rename(columns={'id': 'user_id'}, inplace=True)
bot_df = pd.merge(bot_users, porn_df, on='user_id')

#Propaganda
prop_1 = pd.merge(prop_tweet_intrad, prop_url_intrad, on='user_id')
prop_df = pd.merge(prop_1, prop_tweet_descr, on='user_id')

gen_users.rename(columns={'id': 'user_id'}, inplace=True)
gen_df = pd.merge(gen_users, prop_df, on='user_id')

# Merge users from old train

gen.columns

users = pd.read_csv('data/full/train.csv')

bots = users[users.target != 4].sample(n=2500, random_state=42)
gen = users[users.target == 4].sample(n=2500, random_state=42)

## Drop context score
bots['target'] = 1
gen['target'] = 0

bots.drop(columns=['porn_words_score', 'spam_words_score', 'prop_words_score', 'fake_words_score', 'genuine_words_score'], inplace=True)
gen.drop(columns=['porn_words_score', 'spam_words_score', 'prop_words_score', 'fake_words_score', 'genuine_words_score'], inplace=True)



bot_df = pd.concat([bot_df, bots])

gen_df = pd.concat([gen_df, gen])

bot_df = bot_df.reset_index().drop(columns=['index'])

gen_df = gen_df.reset_index().drop(columns=['index'])

bot_or_not_df = pd.concat([gen_df, bot_df])

bot_or_not_df = bot_or_not_df.reset_index().drop(columns=['index'])

bot_or_not_df.to_csv('data/bot_or_not/train.csv', index=False)

## Controllo integrità

porn_df = pd.read_csv('data/porn/train.csv')
news_df = pd.read_csv('data/propaganda/train.csv')
spam_df = pd.read_csv('data/spam/train.csv')
fake_df = pd.read_csv('data/fake_followers/train.csv')
gen_df = pd.read_csv('data/genuine/train.csv')

len(porn_df)

len(set(porn_df.user_id))

len(fake_df)

len(set(fake_df.user_id))

len(news_df)

len(set(news_df.user_id))

len(spam_df)

len(set(spam_df.user_id))

len(gen_df)

len(set(gen_df.user_id))
