==> Code/classification/25-refill.ipynb <==
import pandas as pd

full = pd.read_csv('../nuovi notebooks/data/bon/full_preproc_train.csv')

len(full.columns)

full.columns

train = pd.read_csv('data/full/train_1000.csv')

train.columns

train = train.drop(columns=['Unnamed: 0', 'nsfw_avg', 'porn_words_score', 'prop_words_score', 'spam_words_score', 'fake_words_score'])

import numpy as np

mask = np.invert(np.in1d(full.columns, train.columns))

full.columns[mask]

full.verified.value_counts()

bots = pd.read_csv('data/bon/bots/users.csv', engine='python', encoding='utf-8-sig')

humans = pd.read_csv('data/bon/humans/users.csv', engine='python', encoding='utf-8-sig')

def find_verfied(user):
    if user.target == 1:
        selection = bots[bots.friends_count==user.friends_count][bots.followers_count == user.followers_count][bots.statuses_count == user.statuses_count][bots.listed_count == user.listed_count]
        if len(selection) == 1:
            return selection.verified
        else:
            return 0
    else:
        selection = humans[humans.friends_count==user.friends_count][humans.followers_count == user.followers_count][humans.statuses_count == user.statuses_count][humans.listed_count == user.listed_count]
        if len(selection) == 1:
            return selection.verified
        else:
            return 0

len(full[full.target==0])

verified = []

for user in full.itertuples():

    if user.target == 1:
        selection = bots[bots.friends_count==user.friends_count][bots.followers_count == user.followers_count]
        if len(selection) == 1:
            if selection.iloc[0].verified == 'True':
                verified.append(1)
            else:
                verified.append(0)
        else:
            verified.append(0)
    else:
        selection = humans[humans.friends_count==user.friends_count][humans.followers_count == user.followers_count]
        if len(selection) == 1:
            if selection.iloc[0].verified == 'True':
                verified.append(1)
            else:
                verified.append(0)
        else:
            verified.append(0)

full['verified'] = verified

full.verified.value_counts()

len(full.columns)

len(train.columns)

train = train.drop('user_id', axis=1)

len(train.columns)

train['verified'] = 0

from collections import Counter
from sklearn.preprocessing import MinMaxScaler, normalize

tweets = pd.concat([pd.read_csv('data/porn/tweets.csv', encoding='utf-8-sig', sep='\t'),
                    pd.read_csv('data/propaganda/tweets.csv', encoding='utf-8-sig', sep='\t'),
                    pd.read_csv('data/spam/tweets.csv', encoding='utf-8-sig', sep='\t'),
                    pd.read_csv('data/fake_followers/tweets.csv', encoding='utf-8-sig', sep='\t'),
                    pd.read_csv('data/genuine/tweets.csv', encoding='utf-8-sig', sep='\t')
                   ])

bot_words = pd.read_csv('data/bon/bots/filtered_main_words.csv', sep=',')
gen_words = pd.read_csv('data/bon/humans/filtered_main_words.csv', sep=',')

refill = pd.concat([
    train[train.target==0].sample(2000),
    train[train.target==1].sample(2000),
    train[train.target==2].sample(2000),
    train[train.target==3].sample(2000)
])

genuine = pd.read_csv('data/full/rf_preproc_train.csv')

genuine = genuine[genuine.target==4]

genuine_entropy = pd.read_csv('data/bon/full/validation.csv')

entropies = []

for user in genuine.itertuples():

    selection = genuine_entropy[genuine_entropy.friends_count==user.friends_count][genuine_entropy.followers_count == user.followers_count]
    if len(selection) == 1:
        entropies.append(selection.iloc[0].url_entropy)
    else:
        entropies.append(0)

genuine['url_entropy'] = entropies

mask = np.in1d(genuine.columns, train.columns)

genuine = genuine[genuine.columns[mask]]

genuine['verified'] = 0

len(refill), len(genuine)

refill = pd.concat([refill, genuine])

refill = refill.reset_index()

def edit_target(x):
    if x<=3:
        return 1
    else:
        return 0
    
refill['target'] = refill['target'].apply(lambda x: edit_target(x))

multiclass_users = pd.read_csv('data/full/rf_preproc_train.csv')

ids = []

for user in refill.itertuples():

    selection = multiclass_users[multiclass_users.friends_count==user.friends_count][multiclass_users.followers_count == user.followers_count][multiclass_users.statuses_count == user.statuses_count]
    if len(selection) == 1:
        ids.append(selection.iloc[0].user_id)
    else:
        ids.append(0)

refill['user_id'] = ids

refill=refill[refill.user_id != 0]

def compute_bot_not_score(tweets):

    user_score = pd.DataFrame(columns=['bot_words_score', 'gen_words_score'])

    for tweet in tweets:
        # check for words in main_words and compute the scores for each tweet and for each category
        mask = np.in1d(bot_words.word, tweet.split())
        bot_score = bot_words.loc[mask]['score'].values.sum()
        mask = np.in1d(gen_words.word, tweet.split())
        gen_score = gen_words.loc[mask]['score'].values.sum()
        
        user_score = user_score.append(pd.DataFrame({'bot_words_score': bot_score, 'gen_words_score': gen_score}, index=[0]), ignore_index=True)

    return user_score

def bot_not_score(tweets_df, id):
    
    tweets = tweets_df[tweets_df.user_id.astype(int) == int(id)]['full_text']
    if len(tweets) > 0:
        # sum all the scores of each category
        user_score = compute_bot_not_score(tweets).sum()
        scores = np.divide(user_score,len(tweets))
    else:
        scores = pd.DataFrame({'bot_words_score': 0, 'gen_words_score': 0}, index=[0])
    
    # return the average scores of each user
    return scores

from IPython.display import clear_output

def context_bot_not_score(bot_ids, tweets_df):
        
    score_df = pd.DataFrame(columns=['bot_words_score', 'gen_words_score', 'user_id'])
    i = 0
    for user_id in bot_ids:
        i += 1
        scores = bot_not_score(tweets_df, user_id)
        score_df = score_df.append(scores, ignore_index=True)
        if (i%10 == 0):
            clear_output()
            print(str(i) +  " bots processed")
    
    score_df['user_id'] = bot_ids.values
    
    score_df.reset_index(drop=True, inplace=True)
    #score_df.to_csv('data/' + bot_type + '/context_score.csv', index=False)
    print("done!")
    return score_df

scores = context_bot_not_score(refill.user_id, tweets)

refill = refill.merge(scores, on='user_id')

refill.to_csv('data/bon/full/train_with_refill.csv', index=False)

## merge per rimediare alla cazzata

refill = pd.read_csv('data/bon/full/train_with_refill.csv')

refill = refill.drop(columns=['user_id', 'index'])

refill = pd.concat([full,refill])
